lat/lon update
Had to convert exif format of DMS to decimal. Added npm sharp library to convert all photos to JPEG (except for HEIC, but it works fine with that format). Added check for image size. Searches all nested objects for lat/lon data. Need to add function to prompt user to input lat/lon info if it's not there.